Move the services api to app start
