+++ app/src/main/java/org/listenbrainz/android/util/ListenSessionListener.kt
@@ -6,20 +6,11 @@ import android.media.session.MediaSessionManager.OnActiveSessionsChangedListener
 import android.media.session.PlaybackState
 import com.dariobrux.kotimer.Timer
 import com.dariobrux.kotimer.interfaces.OnTimerListener
-import org.listenbrainz.android.model.ListenBrainzExternalServices
 import org.listenbrainz.android.repository.AppPreferences
-import org.listenbrainz.android.service.ListensService
 import org.listenbrainz.android.util.Log.d
 import org.listenbrainz.android.util.Log.w
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
-class ListenSessionListener(
-    private val handler: ListenHandler,
-    val appPreferences: AppPreferences,
-    val service: ListensService,
-) : OnActiveSessionsChangedListener {
+class ListenSessionListener(private val handler: ListenHandler, val appPreferences: AppPreferences) : OnActiveSessionsChangedListener {
     
     private val activeSessions: MutableMap<MediaController, ListenCallback?> = HashMap()
 
@@ -27,50 +18,11 @@ class ListenSessionListener(
         d("onActiveSessionsChanged: EXECUTED")
         if (controllers == null) return
         clearSessions()
-
-        val token = appPreferences.lbAccessToken
-        val username = appPreferences.username
-        if (token.isNullOrEmpty()) {
-            d("ListenBrainz User token has not been set!")
-            return
-        }
-
-        if (username != null) {
-            service.getServicesLinkedToAccount("Bearer: $token", username).enqueue(object : Callback<ListenBrainzExternalServices> {
-                override fun onResponse(call: Call<ListenBrainzExternalServices>, response: Response<ListenBrainzExternalServices>) {
-
-                    if (response.isSuccessful) {
-                        d("Services found: " + response.body().toString())
-                        registerControllers(controllers, services = response.body())
-                    } else {
-                        d("Services api failed: " + response.message())
-                        registerControllers(controllers)
-                    }
-                }
-
-                override fun onFailure(call: Call<ListenBrainzExternalServices?>, t: Throwable) {
-                    d("Services not found")
-                    registerControllers(controllers)
-                }
-            })
-        }
-        else {
-            registerControllers(controllers)
-        }
+        registerControllers(controllers)
     }
 
-    fun registerControllers(controllers: List<MediaController>, services: ListenBrainzExternalServices? = null) {
+    private fun registerControllers(controllers: List<MediaController>) {
         for (controller in controllers) {
-            // Check if spotify is already linked with web. Skip if yes.
-            if (controller.packageName == Constants.SPOTIFY_PACKAGE_NAME && services?.services?.contains("spotify") == true) {
-               d("Spotify is already linked with web. Skipping.")
-                if(!appPreferences.listeningBlacklist.contains(controller.packageName)) {
-                    d("Adding Spotify to blacklist.")
-                    appPreferences.listeningBlacklist = appPreferences.listeningBlacklist.plus(controller.packageName)
-                }
-                continue
-            }
-
             // BlackList
             if (controller.packageName in appPreferences.listeningBlacklist)
                 continue
